requirements.txt
flask
paho-mqtt
prometheus_client
python-dotenv
